# apps/api/api/openapi-bundled.yaml
openapi: 3.0.3
info:
  title: Lexicon Backend API
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development
paths:
  /v1/council/sessions:
    post:
      summary: Create new deliberation session
      description: |
        Creates a deliberation session from a case summary.

        By default this endpoint returns JSON after the first initial judge
        message has been generated. When the request includes
        `Accept: text/event-stream`, the same creation flow is streamed as SSE:
        setup progress, session creation, first judge chunks, and a final
        completion event. The existing
        `/v1/council/deliberation/{session_id}/stream/initial` endpoint can then
        stream the remaining initial judges.
      operationId: createCouncilSession
      tags:
        - council
      security:
        - TurnstileToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              case_summary: Terdakwa Budi Santoso, PNS Golongan IV/a pada Dinas PUPR Kabupaten X, didakwa telah melakukan tindak pidana korupsi terkait proyek pembangunan jalan senilai Rp 15 miliar. Jaksa mendakwa terdakwa telah menerima suap sebesar Rp 500 juta dari kontraktor PT Maju Jaya untuk memenangkan tender proyek tersebut. Terdakwa merupakan pelaku pertama kali dengan masa kerja 20 tahun. Kerugian negara yang terbukti adalah Rp 2,5 miliar akibat mark-up harga dan kualitas material yang tidak sesuai spesifikasi.
              case_type: corruption
              structured_data:
                defendant_age: 52
                defendant_first_offender: true
                state_loss_idr: 2500000000
              input_type: text_summary
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
              example:
                session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                parsed_case:
                  case_type: corruption
                  summary: Kasus korupsi proyek pembangunan jalan senilai Rp 15 miliar
                  defendant_profile:
                    is_first_offender: true
                    age: 52
                    occupation: PNS Golongan IV/a Dinas PUPR
                  key_facts:
                    - Menerima suap Rp 500 juta dari kontraktor
                    - Mark-up harga proyek
                    - Kualitas material tidak sesuai spesifikasi
                  charges:
                    - Pasal 12 huruf a UU 31/1999 jo UU 20/2001
                    - Pasal 3 UU 31/1999
                  corruption:
                    state_loss_idr: 2500000000
                    position: Kepala Dinas PUPR
                similar_cases: []
                initial_message:
                  id: msg_001
                  session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                  sender:
                    type: agent
                    agent_id: strict
                  content: Berdasarkan fakta hukum yang terungkap, terdakwa jelas telah melanggar Pasal 12 huruf a UU Tipikor...
                  cited_cases:
                    - 01HN6ABC123
                  cited_laws:
                    - Pasal 12 huruf a UU 31/1999
                  timestamp: '2024-03-15T10:30:00Z'
            text/event-stream:
              schema:
                $ref: '#/components/schemas/CreateSessionStreamEventData'
              example:
                event_type: session_created
                session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                parsed_case:
                  case_type: corruption
                  summary: Kasus korupsi proyek pembangunan jalan senilai Rp 15 miliar
                similar_cases: []
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: case_summary must be at least 50 characters
                status_code: 400
    get:
      summary: List deliberation sessions
      operationId: listCouncilSessions
      tags:
        - council
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - active
              - concluded
              - archived
          description: Filter by session status
          example: active
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          example: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          example: 0
      responses:
        '200':
          description: Sessions list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSessionsResponse'
              example:
                sessions:
                  - id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                    status: active
                    case_input:
                      input_type: text_summary
                      raw_input: Terdakwa Budi Santoso...
                      parsed_case:
                        case_type: corruption
                        summary: Kasus korupsi proyek pembangunan jalan
                    similar_cases: []
                    messages: []
                    created_at: '2024-03-15T10:30:00Z'
                    updated_at: '2024-03-15T10:30:00Z'
                  - id: sess_01HN6VJKM4XPQW3B5RTCDEFGHJ
                    status: concluded
                    case_input:
                      input_type: text_summary
                      raw_input: Terdakwa Ahmad Hidayat...
                      parsed_case:
                        case_type: narcotics
                        summary: Kasus kepemilikan narkotika jenis sabu
                    similar_cases: []
                    messages: []
                    created_at: '2024-03-14T09:00:00Z'
                    updated_at: '2024-03-14T15:30:00Z'
                    concluded_at: '2024-03-14T15:30:00Z'
                pagination:
                  limit: 20
                  offset: 0
                  total: 25
  /v1/council/sessions/{session_id}:
    get:
      summary: Get session details
      operationId: getCouncilSession
      tags:
        - council
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
          example: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSessionResponse'
              example:
                session:
                  id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                  status: active
                  case_input:
                    input_type: text_summary
                    raw_input: Terdakwa Budi Santoso, PNS Golongan IV/a pada Dinas PUPR...
                    parsed_case:
                      case_type: corruption
                      summary: Kasus korupsi proyek pembangunan jalan senilai Rp 15 miliar
                      defendant_profile:
                        is_first_offender: true
                        age: 52
                        occupation: PNS Golongan IV/a Dinas PUPR
                      key_facts:
                        - Menerima suap Rp 500 juta dari kontraktor
                        - Kerugian negara Rp 2,5 miliar
                      charges:
                        - Pasal 12 huruf a UU 31/1999
                      corruption:
                        state_loss_idr: 2500000000
                        position: Kepala Dinas PUPR
                  similar_cases:
                    - case_id: 01HN6ABC123
                      case_number: 45/Pid.Sus-TPK/2023/PN.Jkt.Pst
                      similarity_score: 0.87
                      similarity_reason: Kasus korupsi proyek infrastruktur dengan modus mark-up serupa
                      verdict_summary: Terdakwa terbukti bersalah, divonis 6 tahun penjara
                      sentence_months: 72
                  messages:
                    - id: msg_001
                      session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                      sender:
                        type: agent
                        agent_id: strict
                      content: Berdasarkan fakta hukum, terdakwa jelas melanggar Pasal 12 huruf a...
                      cited_cases:
                        - 01HN6ABC123
                      cited_laws:
                        - Pasal 12 huruf a UU 31/1999
                      timestamp: '2024-03-15T10:30:00Z'
                  legal_opinion: null
                  created_at: '2024-03-15T10:30:00Z'
                  updated_at: '2024-03-15T10:35:00Z'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: session not found
                status_code: 404
    delete:
      summary: Delete session
      operationId: deleteCouncilSession
      tags:
        - council
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
          example: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: session deleted successfully
  /v1/council/sessions/{session_id}/conclude:
    post:
      summary: Conclude deliberation session
      operationId: concludeCouncilSession
      tags:
        - council
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
          example: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
      responses:
        '200':
          description: Session concluded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSessionResponse'
              example:
                session:
                  id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                  status: concluded
                  case_input:
                    input_type: text_summary
                    raw_input: Terdakwa Budi Santoso...
                    parsed_case:
                      case_type: corruption
                      summary: Kasus korupsi proyek pembangunan jalan
                  similar_cases: []
                  messages: []
                  legal_opinion: null
                  created_at: '2024-03-15T10:30:00Z'
                  updated_at: '2024-03-15T12:00:00Z'
                  concluded_at: '2024-03-15T12:00:00Z'
  /v1/council/deliberation/{session_id}/message:
    post:
      summary: Send message to council
      operationId: sendCouncilMessage
      tags:
        - council
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
          example: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            example:
              content: Bagaimana pendapat Hakim Humanis mengenai faktor-faktor yang meringankan terdakwa, mengingat ini adalah pelanggaran pertamanya?
              target_agent: humanis
      responses:
        '200':
          description: Message sent and response received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
              example:
                user_message:
                  id: msg_002
                  session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                  sender:
                    type: user
                  content: Bagaimana pendapat Hakim Humanis mengenai faktor-faktor yang meringankan terdakwa?
                  intent: ask_opinion
                  cited_cases: []
                  cited_laws: []
                  timestamp: '2024-03-15T10:35:00Z'
                agent_responses:
                  - id: msg_003
                    session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                    sender:
                      type: agent
                      agent_id: humanist
                    content: Terima kasih atas pertanyaannya. Dalam mempertimbangkan faktor-faktor yang meringankan, kita perlu melihat bahwa terdakwa adalah pelaku pertama kali dengan rekam jejak pelayanan publik selama 20 tahun. Hal ini menunjukkan bahwa tindakan korupsi ini mungkin bukan cerminan karakter sejatinya...
                    intent: null
                    cited_cases:
                      - 01HN6DEF456
                    cited_laws:
                      - Pasal 53 KUHP
                    timestamp: '2024-03-15T10:35:15Z'
  /v1/council/deliberation/{session_id}/continue:
    post:
      summary: Continue deliberation
      operationId: continueCouncilDiscussion
      tags:
        - council
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
          example: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinueDiscussionRequest'
            example:
              num_rounds: 2
      responses:
        '200':
          description: Deliberation continued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinueDiscussionResponse'
              example:
                new_messages:
                  - id: msg_004
                    session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                    sender:
                      type: agent
                      agent_id: strict
                    content: Saya tidak sependapat dengan Hakim Humanis. Meskipun terdakwa adalah pelaku pertama kali, kerugian negara yang ditimbulkan sangat besar...
                    cited_cases: []
                    cited_laws:
                      - Pasal 2 ayat (1) UU 31/1999
                    timestamp: '2024-03-15T10:40:00Z'
                  - id: msg_005
                    session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                    sender:
                      type: agent
                      agent_id: historian
                    content: Jika kita melihat preseden dalam kasus serupa, putusan MA No. 1261 K/Pid.Sus/2015 memberikan panduan bahwa...
                    cited_cases:
                      - 01HN6GHI789
                    cited_laws: []
                    timestamp: '2024-03-15T10:40:30Z'
                total_messages: 8
  /v1/council/deliberation/{session_id}/messages:
    get:
      summary: Get deliberation messages
      operationId: getCouncilMessages
      tags:
        - council
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
          example: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
          example: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          example: 0
      responses:
        '200':
          description: Messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMessagesResponse'
              example:
                messages:
                  - id: msg_001
                    session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                    sender:
                      type: agent
                      agent_id: strict
                    content: Berdasarkan fakta hukum yang terungkap, terdakwa jelas telah melanggar Pasal 12 huruf a UU Tipikor...
                    cited_cases:
                      - 01HN6ABC123
                    cited_laws:
                      - Pasal 12 huruf a UU 31/1999
                    timestamp: '2024-03-15T10:30:00Z'
                  - id: msg_002
                    session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                    sender:
                      type: user
                    content: Bagaimana pendapat Hakim Humanis mengenai faktor yang meringankan?
                    intent: ask_opinion
                    cited_cases: []
                    cited_laws: []
                    timestamp: '2024-03-15T10:35:00Z'
                  - id: msg_003
                    session_id: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
                    sender:
                      type: agent
                      agent_id: humanist
                    content: Dalam mempertimbangkan faktor meringankan, terdakwa adalah pelaku pertama kali...
                    cited_cases:
                      - 01HN6DEF456
                    cited_laws:
                      - Pasal 53 KUHP
                    timestamp: '2024-03-15T10:35:15Z'
...
  /v1/council/deliberation/{session_id}/stream/initial:
    post:
      summary: Stream remaining initial opinions from judges
      description: |
        Session creation generates 1 random judge's opinion. This endpoint
        streams opinions from the remaining judges to complete the initial
        deliberation round.
      operationId: streamCouncilInitial
      tags:
        - council
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
          example: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
      responses:
        '200':
          description: SSE stream of deliberation
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/StreamEventData'
              example:
                event_type: agent_start
                agent_id: humanist
  /v1/council/deliberation/{session_id}/stream/message:
    post:
      summary: Stream message response
      description: |
        Send a message and stream agent responses in real-time.
        The orchestrator determines which agent(s) should respond based on
        target_agent, message intent, and conversation balance.
      operationId: streamCouncilMessage
      tags:
        - council
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
          example: sess_01HN6VJKM4XPQW3B5RTCDEFGHN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamMessageRequest'
            example:
              content: Hakim Legalis, apa pendapat Anda tentang vonis yang tepat untuk kasus ini?
              target_agent: legalis
      responses:
        '200':
          description: SSE stream of response
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/StreamEventData'
              example:
                event_type: user_message
                message_id: msg_004
                content: Hakim Strict, apa pendapat Anda...
...
components:
  schemas:
...
    CreateSessionRequest:
      type: object
      required:
        - case_summary
      properties:
        case_summary:
          type: string
          minLength: 50
          maxLength: 10000
          description: Summary of the case for deliberation
        case_type:
          $ref: '#/components/schemas/CouncilCaseType'
        structured_data:
          $ref: '#/components/schemas/StructuredCaseData'
        input_type:
          $ref: '#/components/schemas/InputType'
    CreateSessionResponse:
      type: object
      description: Response after creating a session
      required:
        - session_id
        - parsed_case
        - similar_cases
      properties:
        session_id:
          type: string
        parsed_case:
          $ref: '#/components/schemas/ParsedCaseInput'
        similar_cases:
          type: array
          items:
            $ref: '#/components/schemas/CouncilSimilarCase'
        initial_message:
          $ref: '#/components/schemas/DeliberationMessage'
          nullable: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        status_code:
          type: integer
...
    InputType:
      type: string
      description: Type of case input
      enum:
        - text_summary
        - extraction_id
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Lexicon-Api-Key
      description: API key for bulk data egress (ProcureLens ETL)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentik OIDC JWT token for admin authentication
    CookieAuth:
      type: apiKey
      in: cookie
      name: admin_session
      description: |
        Opaque admin BFF session cookie issued by `/v1/admin/auth/callback`.
        HttpOnly, Secure, SameSite=Lax. Admin endpoints accept EITHER this
        cookie OR `Authorization: Bearer <jwt>`; supplying both causes the
        Bearer token to win.
    TurnstileToken:
      type: apiKey
      in: header
      name: X-Turnstile-Token
      description: Cloudflare Turnstile verification token from client-side widget